修复PowerPoint模块问题： 1. 修复幻灯片版式选项为标准11个版式 2. 删除PPT页脚文字配置中多余的中文文本值参数 3. 为动画计时与延时配置添加指定对象参数 4. 修改操作点分数支持小数点（如0.3分）

+++ ExamLab/Models/OperationPoint.cs
@@ -301,7 +301,7 @@ public class OperationPoint : ReactiveObject
     /// <summary>
     /// 操作点分值
     /// </summary>
-    [Reactive] public int Score { get; set; }
+    [Reactive] public decimal Score { get; set; }
 
     /// <summary>
     /// 关联的评分题目ID
